Clean up Build:
Eliminated the need for api.jar, fedora-client.jar and lims-util.jar by eliminating edu.umd.lib.fedora.sweep.CollStats.java which is no longer used.
So we do not need to retain links to any of these 3 jars for this project any more.
Altered tests to reduce the amount of information that they printed out.
Altered sample files for tests from file system calls to resource calls.  Added the sample files to the test resources.
This all works together to get a clean build of fedora-utilities.